requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
duckdb==0.7.1
astro-sdk-python[duckdb]==1.5.3
pandas==1.5.3
fsspec==2023.3.0
s3fs==2023.3.0